+++ ft-basic/ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: earnaud <earnaud@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/03 18:07:03 by earnaud           #+#    #+#             */
-/*   Updated: 2020/11/03 18:07:05 by earnaud          ###   ########.fr       */
+/*   Updated: 2020/11/07 19:31:06 by earnaud          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (src[i] && i < size)
